Added client methods to user entities
