+++ src/Disqord.Rest/Entities/Core/Users/ICurrentUser.Client.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using System.Threading.Tasks;
+
+namespace Disqord
+{
+    public partial interface ICurrentUser : IUser
+    {
+        Task ModifyAsync(Action<ModifyCurrentUserProperties> action, RestRequestOptions options = null);
+    }
+}
