+++ src/Disqord.Rest/Entities/Core/Users/IMember.cs
@@ -1,10 +1,9 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Threading.Tasks;
 
 namespace Disqord
 {
-    public interface IMember : IUser
+    public partial interface IMember : IUser
     {
         string Nick { get; }
 
@@ -21,17 +20,5 @@ namespace Disqord
         DateTimeOffset? BoostedAt { get; }
 
         bool IsBoosting { get; }
-
-        Task ModifyAsync(Action<ModifyMemberProperties> action, RestRequestOptions options = null);
-
-        Task KickAsync(RestRequestOptions options = null);
-
-        Task BanAsync(string reason = null, int? messageDeleteDays = null, RestRequestOptions options = null);
-
-        Task UnbanAsync(RestRequestOptions options = null);
-
-        Task GrantRoleAsync(Snowflake roleId, RestRequestOptions options = null);
-
-        Task RevokeRoleAsync(Snowflake roleId, RestRequestOptions options = null);
     }
 }
