+++ src/Disqord.Rest/Entities/Rest/Users/RestUser.cs
@@ -4,7 +4,7 @@ using Disqord.Models;
 
 namespace Disqord.Rest
 {
-    public class RestUser : RestSnowflakeEntity, IUser
+    public partial class RestUser : RestSnowflakeEntity, IUser
     {
         public string Name { get; private set; }
 
@@ -33,31 +33,7 @@ namespace Disqord.Rest
 
         public string GetAvatarUrl(ImageFormat? imageFormat = null, int size = 2048)
             => Discord.Internal.GetAvatarUrl(this);
-
-        public Task SetNoteAsync(string note, RestRequestOptions options = null)
-            => Client.SetNoteAsync(Id, note, options);
-
-        public Task<RestDmChannel> CreateDmChannelAsync(RestRequestOptions options = null)
-            => Client.CreateDmChannelAsync(Id, options);
-
-        public async Task<RestUserMessage> SendMessageAsync(string content = null, bool textToSpeech = false, Embed embed = null, RestRequestOptions options = null)
-        {
-            var channel = await CreateDmChannelAsync(options).ConfigureAwait(false);
-            return await channel.SendMessageAsync(content, textToSpeech, embed, options).ConfigureAwait(false);
-        }
-
-        public async Task<RestUserMessage> SendMessageAsync(LocalAttachment attachment, string content = null, bool textToSpeech = false, Embed embed = null, RestRequestOptions options = null)
-        {
-            var channel = await CreateDmChannelAsync(options).ConfigureAwait(false);
-            return await channel.SendMessageAsync(attachment, content, textToSpeech, embed, options).ConfigureAwait(false);
-        }
-
-        public async Task<RestUserMessage> SendMessageAsync(IEnumerable<LocalAttachment> attachments, string content = null, bool textToSpeech = false, Embed embed = null, RestRequestOptions options = null)
-        {
-            var channel = await CreateDmChannelAsync(options).ConfigureAwait(false);
-            return await channel.SendMessageAsync(attachments, content, textToSpeech, embed, options).ConfigureAwait(false);
-        }
-
+        
         public override string ToString()
             => Tag;
     }
