+++ src/Disqord/Entities/Users/CachedMember.cs
@@ -2,7 +2,6 @@
 using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Linq;
-using System.Threading.Tasks;
 using Disqord.Logging;
 using Disqord.Models;
 using Disqord.Models.Dispatches;
@@ -10,7 +9,7 @@ using Qommon.Collections;
 
 namespace Disqord
 {
-    public sealed class CachedMember : CachedUser, IMember
+    public sealed partial class CachedMember : CachedUser, IMember
     {
         public override string Name => SharedUser.Name;
 
@@ -210,23 +209,5 @@ namespace Disqord
 
         public ChannelPermissions GetPermissionsFor(IGuildChannel channel)
             => Discord.Permissions.CalculatePermissions(Guild, channel, this, Roles.Values);
-
-        public Task ModifyAsync(Action<ModifyMemberProperties> action, RestRequestOptions options = null)
-            => Client.ModifyMemberAsync(Guild.Id, Id, action, options);
-
-        public Task KickAsync(RestRequestOptions options = null)
-            => Client.KickMemberAsync(Guild.Id, Id, options);
-
-        public Task BanAsync(string reason = null, int? messageDeleteDays = null, RestRequestOptions options = null)
-            => Client.BanMemberAsync(Guild.Id, Id, reason, messageDeleteDays, options);
-
-        public Task UnbanAsync(RestRequestOptions options = null)
-            => Client.UnbanMemberAsync(Guild.Id, Id, options);
-
-        public Task GrantRoleAsync(Snowflake roleId, RestRequestOptions options = null)
-            => Client.GrantRoleAsync(Guild.Id, Id, roleId, options);
-
-        public Task RevokeRoleAsync(Snowflake roleId, RestRequestOptions options = null)
-            => Client.RevokeRoleAsync(Guild.Id, Id, roleId, options);
     }
 }
